+++ src/test/java/pages/base/RegistrationPage.java
@@ -23,6 +23,9 @@ public class RegistrationPage {
     @FindBy(id = "username")
     public WebElement userName;
 
+    @FindBy(xpath = "//*[@id='register-form']/div[4]/div")
+    public WebElement usernameisrequired;
+
     @FindBy(id = "email")
     public WebElement email;
