+++ src/test/resources/features/ui_features/homePageFeatures.feature
@@ -2,3 +2,4 @@ Feature: HomePage related features
   Background:
     Given user goes to URL "https://www.medunna.com/"
     Then verify the page title contains "MEDUNNA"
+
